bots: join and consume remote streams

+++ bots/index.ts
@@ -1,22 +1,49 @@
-import { createStream, createUser, stopStream } from "./src/procedures";
+import {
+  createStream,
+  createUser,
+  joinStream,
+  stopStream,
+} from "./src/procedures";
 import { initMediasoupWorker } from "./src/media";
-import * as path from "path";
+
+const runStreamer = async () => {
+  const clientRecord = await createUser();
+  await createStream(clientRecord);
+
+  setTimeout(async () => {
+    await stopStream(clientRecord);
+    await clientRecord.api.user.delete();
+
+    clientRecord.api.close();
+  }, 60000);
+};
+
+const runViewer = async () => {
+  const clientRecord = await createUser();
+
+  const { feed } = await clientRecord.api.feed.get(0);
+
+  if (feed.length === 0) {
+    console.log("No streams in feed, exiting");
+    await clientRecord.api.user.delete();
+
+    return;
+  }
+
+  const streamToWatch = feed[0];
+
+  console.log(`joining stream ${streamToWatch.id}`);
+  await joinStream(streamToWatch.id, clientRecord);
+};
 
 const main = async () => {
   await initMediasoupWorker();
 
-  setTimeout(async () => {
-    const clientRecord = await createUser();
-    await createStream(clientRecord);
+  runStreamer();
 
-    console.log("sending media");
-    setTimeout(async () => {
-      await stopStream(clientRecord);
-      await clientRecord.api.user.delete();
-
-      clientRecord.api.close();
-    }, 60000);
-  }, 1000);
+  setTimeout(() => {
+    runViewer();
+  }, 2000);
 };
 
 main();
